1. New terms: long cell and short cell (is_a epidermal cell)
(i) epidermal cell
	(i) long cell (PO:0004717)
	(i) short cell (PO:0004718)
		(i) epidermal cork cell
		(i) silica cell

Definitions:
long cell - a more or less rectangular epidermal cell in which the proximo-distal dimension is several times longer than the transverse dimension.

short cell - an isodiametric epidermal cell.

To make it consistent with its sibling, 'epidermal cork cell', I modified the definition of silica cell to:
"Cells with deposits of silica in them. One of the two types of short cells in the epidermis of grasses and bamboos. It is usually paired with epidermal cork cell."

Comment (for both cell types): In the leaf epidermis of Poaceae and Joinvilleaceae, long cells and short cells often alternate in longitudinal rows.

2. New term - hilum groove
Def: longitudinal groove in the testa through the prominent hilum of the seed.
Synonym: faboid split

3. Stem cell terms
Added stem cell as synonym of 'initial cell, and also, added 'columella root cap stem cell' as synonym of 'columella root cap initial'

4. New term - paraclade and paraclade cortex
Stem
 (p) paraclade (PO:0004720)
	(p) paraclade cortex (PO:0004721)

paraclade (PO:0004720)
Def: lateral branch which ends in an inflorescence that repeats the main inflorescence (used both refs: APWeb and Weberling, ISBN: 0521251346
Morphology of Flowers and Inflorescences.

paraclade cortex (PO:0004721
Def The primary tissue located between the vascular system and the epidermis of the paraclade.

5. Inflorescence stem added as synonym of 'peduncle'.

6. New terms - pedicel cortex and sepal vascular system,

pedicel cortex PO:0004722
Def: The primary tissue located between the vascular system and the epidermis of the pedicel.

sepal vascular system, PO:0004723
Def: Vascular system of the sepal.

7. New term - hypocotyl-root junction PO:0004724

Def: Junction point of the embryonic root and hypocotyl.
Note: the decision was to add this term under 'embryo axis' and also propagate it under shoot.

8. New terms: abaxial and adaxial leaf primorduim

abaxial side of leaf primorduim PO:0004725
Def: The abaxial/outer side of the leaf primordium that becomes the abaxial/lower leaf blade

adaxial side of leaf primorduim PO:0004726
Def: The adaxial/inner side of the leaf primordium that becomes adaxial/upper leaf blade.