plant embryo (PO:0009009) new def'n: A whole plant in the early part of a sporophytic phase after the first cell division.
Revised comment: An embryo is generally formed after the first division
of a zygote, but in the case of adventitious embryos, somatic embryos,
other embryos that arise through apogamy, and cultured haploid embryos,
it is formed after the division of a single cell that is not a zygote.
The end of the embryonic phase varies among taxa. In seed plants, the
embryonic phase ends with germination. In pteridophytes, the embryonic
phase ends with the formation of the first true leaf after the
cotyledon(s). In bryophytes, the embryonic phase ends when the apical
cell stops dividing and the sporangium begins to develop. In cultured
plant embryos, the embryonic phase ends when organs (roots, shoot axes,
or leaves) begin to form.

 cultured embryo (PO:0000010)
Renamed: cultured plant embryo (PO:0000010)

New def’n: A plant embryo that is grown or maintained in vitro.

new Comment: Includes isolated and maintained mature or immature zygotic
embryos, somatic embryos or haploid embryos (derived from male
gametophyte).

cultured somatic embryo (PO:0000011), renamed cultured somatic plant
embryo

revised def'n: A somatic plant embryo arising from previously
differentiated somatic cells and grown and maintained in vitro.

is_a somatic plant embryo and also a cultured plant embryo.