revised def'n., cuticle (PO:0000022):''' A plant anatomical substance composed of a polyester matrix of hydroxy- and hydroxy epoxy fatty acids C16 and C18 long (cutin), embedded and overlayed with cuticular wax and forming a layer on the surface of a plant epidermis.
comment: The cuticle is produced in and secreted by epidermal cells of leaves, stems, fruits, and other aerial plant structures. It has as parts an inner cuticle proper and an outer layer of epicuticular wax.

is_a portion of plant substance, adjacent_to epidermis

ref: PMID:12467640


'''proposed new term: cuticle proper (PO:0025387):''' A portion of plant substance that is the inner layer of a cuticle, composed primarily of cutin, embedded and overlayed with cuticular wax.

synonyms: cutin matrix

ref: PMID:12467640


'''proposed new term: cutin (PO:0025385):''' A portion of plant substance that is composed of a polyester matrix of hydroxy- and hydroxy epoxy fatty acids C16 and C18 long  (''Need to check on other components; may want to work on wording a little'')

Comment: Cutin is the insoluble, covalently cross-linked polymer that forms an electron dense layer over the epidermal cells of an areal plant structure. 

ref: PMID:12467640


'''proposed new term: cuticular wax (PO:0025386):'''  A portion of plant substance comprised predominantly of very long chain aliphatic lipids and is part of a cuticle.

Comment: Also includes triterpenoids and minor secondary metabolites, such as sterols and flavonoids. May be embedded within or external to the layer of cutin of the cuticle proper.  Phenotypically wax distinguished as the glaucous or whitish bloom on a part of a plant shoot system that corresponds to crystalloid epicuticular wax with light scattering capacity.

ref: PMID:12467640, PO_REF:00010


'''Revised def'n for requested term, epicuticular wax(PO:0025388):'''  A portion of cuticular wax that forms crystalloids or a smooth film on the exterior of a cuticle. 

comment: Epicuticular waxes may take a variety of forms, such as plates, ribbons, tubes, and rods.

adjacent_to cuticle proper

ref: PMID:12467640, PO_REF:00010


'''proposed  new term: intercuticular wax (PO:0025389):''' A portion of cuticular wax that is an amorphous mixture of lipids embedded in the cutin of a cuticle proper.

part_of cuticle proper

ref: PMID:12467640, PO_REF:00010