renamed: plant cuticle: PO:0000022 modified definition: A plant anatomical substance composed of a polyester matrix of hydroxy- and hydroxy epoxy fatty acids C16 and C18 chains (cutin), embedded and overlayed with cuticular wax and forming a layer on the surface of a plant epidermis..
revised comment: 
The plant cuticle is produced in and secreted by epidermal cells (PO:0004013) of leaves (PO:0025034), stems (PO:0009047), fruits (PO:0009001), and other aerial plant structures. It has as parts an inner cuticle proper (PO:0025387) and an outer layer of epicuticular wax (PO:0025388).

new name: plant cuticle proper PO:0025387: A portion of plant substance that is the inner layer of a cuticle, composed of a polyester matrix of hydroxy- and hydroxy epoxy fatty acid C16 and C18 chains.
Comment: Cutin is the insoluble, covalently cross-linked polymer that forms an electron dense layer over the epidermal cells of an aerial plant structure.

cuticular wax (PO:0025386): A portion of plant substance comprised predominantly of very long chain aliphatic lipids and is part of a cuticle.

Comment: Also includes triterpenoids and minor secondary metabolites, such as sterols and flavonoids. May be embedded within or external to the layer of cutin of the cuticle proper. Phenotypically, wax is distinguished as the glaucous or whitish bloom on a part of a plant shoot system that corresponds to crystalloid epicuticular wax with light scattering capacity.


 epicuticular wax (PO:0025388): A portion of cuticular wax that forms crystalloids or a smooth film on the exterior of a cuticle.
comment: Epicuticular waxes may take a variety of forms, such as plates, ribbons, tubes, and rods.
adjacent_to cuticle proper

New term: intercuticular wax (PO:0025389): A portion of cuticular wax that is an amorphous mixture of lipids embedded in the cutin of a cuticle proper.
part_of cuticle proper,
comment: The intercuticular wax is located adjacent to the cell wall.