Added new meristem terms discussed 1-10-12: Meristem terms: phyllome adaxial meristem:PO:0025400 A portion of meristem tissue that is found on the adaxial side of a young phyllome (PO:0006001). Comment: Contributes to the thickness of a phyllome (PO:0006001). There are two types of adaxial meristems. In phyllomes with a petiole (PO:0020038) and lamina (PO:0025060), an adaxial rounding meristem occurs later in development, is separated from the apical meristem (PO:0020144), and contributes to the thickness of the petiole and midrib. In unifacial phyllomes, the an adaxial meristem occurs earlier in development, is close to but separate from the apical meristem, and contributes to the thickness and cross-sectional shape of the entire phyllome. If annotating to a leaf, use the more specific term leaf adaxial meristem (PO:0025401).
narrow synonyms: 
phyllome adaxial rounding meristem- needs dbxref: poc curators for now
phyllome adaxial unifacial meristem - "

vascular leaf adaxial meristem: PO:0025401
A portion of phyllome adaxial meristem tissue that is found on the adaxial side of a young vascular leaf (PO:0009025).
Comment:  Contributes to the thickness of a vascular leaf (PO:0009025). There are two types of adaxial meristems. In leaves with a petiole (PO:0020038) and lamina (PO:0025060), such as Arabidopsis thaliana, an adaxial rounding meristem occurs later in development, is separated from the apical meristem (PO:0020144), and contributes to the thickness of th petiole and midrib. In unifacial leaves, the an adaxial unifacial meristem occurs earlier in development, is close to but separate from the apical meristem, and contributes to the thickness and cross-sectional shape of the entire leaf.

 is_a phyllome adaxial meristem, 
part_of vascular leaf, participates_in leaf expansion stage
(PO:0001052)

Note: Need to add the participates in relation once the leaf expansion terms have been done in the PSDS.