Added the following terms based on POC conf call 1-10-12: phyllome abaxial meristem:PO:0025402 A portion of meristem tissue that is found on the abaxial side of a developing phyllome (PO:0006001). Comment: Contributes to the thickness of a phyllome (PO:0006001). There are two types of abaxial meristems. In phyllomes with a petiole (PO:0020038) and lamina (PO:0025060), an abaxial rounding meristem occurs later in development, is separated from the apical meristem (PO:0020144), and contributes to the thickness of the petiole and midrib. In unifacial phyllomes, the abaxial meristem occurs earlier in development, is close to but separate from the apical meristem, and contributes to the thickness and cross-sectional shape of the entire phyllome. If annotating to a leaf, use the more specific term leaf abaxial meristem (PO:0025403).
narrow synonyms: 
phyllome abaxial rounding meristem- poc:curators
phyllome abaxial unifacial meristem - poc curators for now

leaf abaxial meristem: PO:0025403
A portion of phyllome abaxial meristem tissue that is found on the abaxial side of a developing vascular leaf.
Comment:  Contributes to the thickness of a vascular leaf (PO:0009025). There are two types of abaxial meristems. In leaves with a petiole (PO:0020038) and lamina (PO:0025060), such as Arabidopsis thaliana, an abaxial rounding meristem occurs later in development, is separated from the apical meristem (PO:0020144), and contributes to the thickness of the petiole and midrib. In unifacial leaves, the abaxial unifacial meristem occurs earlier in development, is close to but separate from the apical meristem, and contributes to the thickness and cross-sectional shape of the entire leaf.

narrow synonyms: 
leaf abaxial rounding meristem- poc:curators
leaf abaxial unifacial meristem - poc:curators