Rhizoid: removed ISBN:Unknown as it created a link to a mass market paperback "Unknown" Added PO:id to definitions of gamete;
coleoptile emergence stages- and coleorhiza emergence stages: 
Modified comment: Occurs in Poaceae and some other monocotyledons. May occur simultaneously with coleoptile emergence from the growth medium.  The actual point of emergence from the seed coat (PO:0009088) may not be observed if the seed (PO:0009010) is underneath a growth medium. In a fruit (PO:0009001) with a persistent pericarp (PO:0009084), emergence from the seed coat may not be observed. If the fruit is a caryopsis, the dry pericarp is fused with the seed coat, and emergence from the pericarp is simultaneous with emergence from the seed coat. This terms is used only for seed plants. 


cotyledon emergence stage (PO:0007049); epicotyl emergence stage (PO:0007054); hypocotyl emergence stage (PO:0007043); radicle emergence stage (PO:0007015)
Added to following to the comment of each: 
In a fruit (PO:0009001) with a persistent pericarp (PO:0009084), emergence from the seed coat may not be observed. If the fruit is a caryopsis, the dry pericarp is fused with the seed coat, and emergence from the pericarp is simultaneous with emergence from the seed coat. This terms is used only for seed plants.