Modified comment of tubercle (PO:0025352): An enlarged leaf base (PO:0020040) that is fused with adjacent shoot axis (PO:0025029) tissue. Currently: "In species with tubercles, the areole (PO:xxxxxxx) is often found at.... etc
We do not have the term 'areole'- replaced it with the new-ish term areole bud
(PO:0025353)
Revised comment:
"In species with tubercles, the areole bud (PO:0025353) is often found at the distal end of the tubercle. The areole bud originates in the leaf axil (PO:0009023) of the leaf that becomes the tubercle, but is displaced to the tip through elongation of the tubercle."

added po:ids to the definitions of tubercle (PO:0025352), leaf base PO:0020040, phyllome base (PO:0025140), leaf (PO:0025034).