New terms added:
plant sap (new PO:00255380): A portion of plant substance (PO:0025161) that is an aqueous solution which may contain sugars, hormones, mineral elements and/or other nutrients.

Comment: plant sap should not be confused with latex or resins.


xylem sap (new PO:0025539): A plant sap (PO:0025538) that is an aqueous solution which may contain mineral elements, nutrients, and plant hormones, and is transported from the root system (PO:0025025) toward the leaves through the tracheary elements (PO:0000290) of the xylem.

Transport of sap in xylem is characterized by movement from the roots toward the leaves.

Phloem sap (new PO:0025539): A plant sap (new PO:0025538) that is an aqueous solution which may contain sugars, mineral elements, and plant hormones, and is transported through the phloem sieve elements (PO:0025406).