reated new terms to describe protoxylem patterns: mesarch protoxylem (new PO:0025553)- A portion of protoxylem (PO:0000272) tissue in which the maturation of primary xylem (PO:0005849) of the shoot system (PO:0009006) progresses both centripedally and centrifugally from the center. comment: Mesarch protoxylem is found in ferns and cycads stems.
exarch protoxylem (PO:0025554): A portion of protoxylem (PO:0000272) tissue in which the maturation of primary xylem (PO:0005849) progresses centripedally, the oldest elements (protoxylem) are farthest from the center of the axis.  Comment: Typical of roots in seed plants. 

shoot system exarch protoxylem (PO:0025556): A portion of exarch protoxylem (PO:002554)tissue in which the maturation of primary xylem (PO:0005849) of the shoot system (PO:0009006) progresses centripedally, the oldest elements (protoxylem) are farthest from the center of the axis.

root exarch protoxylem (PO:0025555): A portion of exarch protoxylem (PO:002554) tissue in which the maturation of primary xylem (PO:0005849) of the root system (PO:0025025) progresses centripedally, the oldest elements (protoxylem) are farthest from the center of the axis.

monarch protoxylem (new PO:0025558)-A portion of root exarch protoxylem (PO:0025555) tissue in which the primary xylem (PO:0005849) of the root system (PO:0025025) hasonly one protoxylem strand or pole, and differentation progresses centripedally, with the oldest element (protoxylem) farthest from the center of the axis.

diarch protoxylem (new PO:0025559)- A portion of root exarch protoxylem (PO:0025555) tissue in which the primary xylem (PO:0005849) of the root system (PO:0025025)  has two protoxylem strands or poles, and differentation progresses centripedally, with the oldest elements (protoxylem) farthest from the center of the axis.

triarch protoxylem (new PO:0025560)- A portion of root exarch protoxylem (PO:0025555) tissue in which the primary xylem (PO:0005849) of the root system (PO:0025025)  has three protoxylem strands or poles, and differentation progresses centripedally, with the oldest elements (protoxylem) farthest from the center of the axis. 

tertrach protoxylem (new PO:0025561)- A portion of root exarch protoxylem (PO:0025555) tissue in which the primary xylem (PO:0005849) of the root system (PO:0025025)  has four protoxylem strands or poles, and differentation progresses centripedally, with the oldest elements (protoxylem) farthest from the center of the axis.

pentarch protoxylem (new PO:0025562)-A portion of root exarch protoxylem (PO:0025555) tissue in which the primary xylem (PO:0005849) of the root system (PO:0025025)  has five protoxylem strands or poles, and differentation progresses centripedally, with the oldest elements (protoxylem) farthest from the center of the axis.
 
polyarch protoxylem (new PO:0025563)- A portion of root exarch protoxylem (PO:0025555) tissue in which the primary xylem (PO:0005849) of the root system (PO:0025025)  has more than five protoxylem strands or poles, and differentation progresses centripedally, with the oldest elements (protoxylem) farthest from the center of the axis.

added dbxrefs OBO_SF_PO:3610599 ; ISBN: 0471245208