Updated trichome terms with the new GO ids for parts of the unicellular trichome: trichome apex (PO:0025550) Updated the comment to read: "This term is the parent to the apex of a multicellular trichome  (PO:0025162) or a unicellular trichome (PO:0025537). An apex of a unicellular trichome is a GO sub-cellular component. Unicellular trichome apex (GO:0090552):  A cell projection part that is the apical most portion of a unicellular trichome. [GOC:PO_curators, PO:0025537]." Added new dbxref to OBO_SF_GO:3611615
multicellular trichome apex (PO:0025551):
Updated the comment to read:  An apex of a unicellular trichome is a GO sub-cellular component. Unicellular trichome apex (GO:0090552):  A cell projection part that is the apical most portion of a unicellular trichome. [GOC:PO_curators, PO:0025537]. Added new dbxref to OBO_SF_GO:3611615

trichome tip PO:0025549: 
Updated the comment to read:
This term is the parent to the tip of a multicellular trichome (PO:0025162) or a unicellular trichome (PO:0025537). A tip of a unicellular trichome is a GO sub-cellular component. Unicellular trichome tip (GO:0090553): A cell projection part that is the apical most portion of a unicellular trichome apex. [GOC:PO_curators]. Added new dbxref to OBO_SF_GO:3611615