Added references to new GO terms on the comment of: unicellular trichome (PO:0025537) for the parts. For parts of the unicellular trichome (such as branch, tip, apex), please see the GO cellular component terms; unicellular trichome apex (GO:0090552),  unicellular trichome tip (GO:0090553), and unicellular trichome branch (GO:0090405)."
Revised comment on multicellular trichome (PO:0000282):
A multicellular trichome (PO:0000282) has as parts more than one trichome cell (PO:0008030), may vary in size and complexity, be branched and may contain a secretory gland.
Revised comment on trichome apex (PO:0025550):
This term is the parent to the apex of a multicellular trichome (PO:0025162) or a unicellular trichome (PO:0025537), specifically, a multicellular trichome apex (PO:0025551) and an unicellular trichome apex (GO:0090552) which is a GO cellular component term.

Revised comment on multicellular trichome apex PO:0025551:
This term describes the apex region of a multicellular trichome (PO:0000282).  The apex of a unicellular trichome is a GO cellular component: unicellular trichome apex (GO:0090552).


Revised comment on trichome tip (PO:0025549):
This term is the parent to the tip of a multicellular trichome (PO:0025162) or the tip of a unicellular trichome (PO:0025537), specifically, a multicellular trichome tip (PO:0025535) and a unicellular trichome tip (GO:0090553), which is a GO cellular component term.

This term describes the apex region of a tip of a multicellular trichome (PO:0000282). The tip may be various shapes.  A tip of a unicellular trichome is a GO cellular component term; unicellular trichome tip (GO:0090553).

Revised comment on multicellular trichome tip (PO:0025535):
This term describes the apex region of a tip of a multicellular trichome (PO:0000282). The tip may be various shapes.  A tip of a unicellular trichome is a GO cellular component term; unicellular trichome tip (GO:0090553).