Revised the following definition: perianth (PO:0009058): A collective phyllome structure (PO:0025023) that includes as parts the corolla (PO:0009059) and/or the calyx (PO:0009060); or one or more tepals (PO:0009033).
Added comment for clarity: The perianth can consist of only the corolla (PO:0009059), only the calyx (PO:0009060) or it may have both structures in a complete flower. If the petals (PO:0009032) and sepals (PO:0009031) are indistinguishable from one another, the perianth consists of just a collective tepal structure (PO:0025021), composed of one or more tepals (PO:0009033), for example in tulip or lily flowers.