Fixed grammar error in calyx (PO:0009060) definition.
Added comment to calyx (PO:0009060): This definition is based on the perfect flower, which has all four whorls. In many species, and in mutants one or more of the whorls are missing. 

Added comment to sepals (PO:0009031): In some species the sepals (PO:0009031) may not be green for example many members of the Liliaceae family.