Minor grammar adjustment in collective tepal structure (PO:0025021) definition.
Added comment to collective tepal structure (PO:0025021): This structure is analogous to the calyx and corolla whorls. 

Changed definition of tepal (PO:0009033): A phyllome (PO:0006001) that is part of a perianth (PO:0009058) where it is not possible to distinguish between the petals (PO:0009032) and sepals (PO:0009031).  

Added comment to tepal (PO:0009033): There is no differentiation between the petals (PO:0009032) and the sepals (PO:0009031) in color, placement, and size.