Changed definition of fused collective tepal structure (PO:0025138) to: A collective tepal structure (PO:0025021) in which  there are at least two fused tepals (PO:0025137).
Add connate tepal structure as a synonym for fused collective tepal structure (PO:0025138). 

Changed definition of free tepal (PO:0025136) to:  A tepal (PO:0009033) that is part of a collective tepal structure (PO:0025021), along with a fused collective tepal structure (PO:0025138), but is separate from (PATO:0001505) the other tepals. 

Free tepal (PO:0025136) is no longer a part of fused collective tepal structue  (PO:0025138). 

Changed definition of fused tepal (PO:0025137) to: A tepal (PO:0009033) that is part of a fused collective tepal structure (PO:0025138) in which it is fused to at least one other tepal.

Changed comment of fused tepal (PO:0021537) to: This is a phenotype that is a cross-product of PO:0025136 (tepal) and PATO:0000642 (fused with).