Changed comment of free teal (PO:0025136): A free tepal (PO:0025136) only occurs in a perianth (PO:0009058) that contains a fused collective tepal structure (PO:0025138). If you are describing a tepal (PO:0025137) that is part of a perianth (PO:0009058) where none of the tepals (PO:0009033) are fused, you should use tepal.
Changed free petal (PO:0025569) definition to: A petal (PO:0009032) that is part of a perianth (PO:0009058) that contains fused sepals (PO:PO:0025584) and/or fused petals (PO:0025583), but is separate from the other sepals (PO:0009031) and petals (PO:0009032).

Add comment to free petal (PO:0025569): A free petal (PO:0025569) only occurs in a perianth (PO:0009058) that contains fused sepals (PO:0025584) and/or fused petals (PO:0025583). If you are describing a petal (PO:0009032) that is part of a perianth (PO:0009058) where none of the sepals (PO:0009031) and/or petals (PO:0009032) are fused, you should use petal.