oot pith (PO:0005040): Revised definition to match exisiting relationship to root stele: was: A portion of pith (PO:0006109) that is part of a root. part_of root stele Revised definition: A portion of pith (PO:0006109) that is part of a root stele (PO:0020124). Added missing xrefs.
root stele (PO:0020124): added po:ids and missing xref.

Added new term for strawberry:
receptacle pith (PO:0025590): A portion of pith (PO:0006109) that is part of a receptacle (PO:0009064). part_of receptacle.
added po:ids

paraclade cortex (PO:0004721):
Revised parentage: moved paraclade cortex (PO:0004721) to is_a cortex (PO:0005708).
revised def'n: A portion cortex (PO:0005708) that is  part of an inflorescence branch (PO:0009081), also called a paraclade.
Added exact synonym: inflorescence branch cortex
updated Xref to sf.

Added new term as a child of cortex:
receptacle cortex (PO:0025591): A portion of cortex (PO:0005708) that is part of a receptacle (PO:0009064).

This will be used to describe the fleshy part of the strawberry.

Revised def'n: embryo cortex (PO:0005014): A portion of cortex (PO:0005708) that is part of a plant embryo (PO:0009009).
Updated the xref

Current definition of pedicel cortex (PO:0004722): A portion of ground tissue that is part of a cortex and part of a pedicel.
Revised def'n: A portion of cortex (PO:0005708) that is part of a pedicel (PO:0009052). Moved from is_a portion of ground tissue PO:0025059

Current definition of root cortex (PO:0000258): A portion of ground tissue (PO:0025059) that is part of a cortex and part of a root (PO:0009005).
Revised to remove redundancy and make it consistent with siblings: A portion of cortex (PO:0005708) that is part of a root (PO:0009005).
Added missing xref


Current definition of petiole cortex (PO:0000053): A portion of ground tissue that is part of a cortex and part of the petiole.
Revised to remove redundancy and make it consistent with siblings: A portion of cortex (PO:0005708) that is part of a petiole (PO:0020038).
Added missing xref

Current definition of shoot axis cortex (PO:0000223): A portion of ground tissue (PO:0025059) that is part of a cortex (PO:0005708) and part of a shoot axis (PO:0025029).

Revised def'n: A portion of cortex (PO:0005708) that is part of a shoot axis (PO:0025029).
Added missing xref