Revised name and def'n:   "shoot node" to "shoot axis node" (PO:0005004) . shoot axis node (PO:0005004): A cardinal organ part (PO:0025001) of a shoot axis (PO:0025029) where a bud (PO:0000055) may form and develop into a branch (PO:0025073), leaf (PO:0025034), flower (PO:0004541), inflorescence (PO:0009049) or nodal root (PO:0003005).
Revised root, since shoot-borne roots are also roots, and to get rid of the circularity with the definition of root system.

Revised def'n: A plant axis (PO:0025004) that is usually positively geotropic and found below ground, lacks shoot axis nodes (PO:0005004) and does not bear leaves (PO:0025034). 

Revised Comment: Roots primarily function in absorption of water and inorganic nutrients, anchoring and supporting the plant body, storage of food and nutrients, and vegetative reproduction.  Roots are endogenous in origin, indeterminate in growth and often with secondary thickening.  Commonly thought of as one of the three basic parts of the seed plant body.