Revised definition: root (PO:0009005): A plant axis (PO:0025004) that lacks shoot axis nodes (PO:0005004), grows indeterminately and is usually positively geotropic.
Revised Comment: Roots function in the absorption of water and inorganic nutrients, anchoring the plant body to the ground, and supporting it, storage of food and nutrients, and vegetative reproduction. The roots of most vascular plant species enter into symbiosis with soil-borne microorganisms. Roots are usually found underground, although there are many exceptions, such as the aerial roots of orchids. Roots often form secondary thickening from the root lateral meristem (PO:0006308). Commonly thought of as one of the three basic parts of the plant body, along with the shoot axis (PO:0025029) and leaves (PO:0025034).

- Removed cross products on shoot lateral meristem (PO:0006344) and root lateral meristem (PO:0006308) as they were redundant.