Moved to be is_a native plant cell and updated the def'n:
brachycyte (PO:0030058): 
A native plant cell (PO:0025606) that develops from a cell in a protonema (PO:0030003) and has a thick cell wall.
Comment: Found in mosses. A brachycyte is a drought tolerant cell that can act as a gemma and may give rise to a new plant. Cells of a protonema (PO:0030003) may develop into brachycytes under stress. May form a chain or be separated by tmema cells (PO:0030059).