Updated Xrefs: - coleorhiza emergence stage (PO:0025475); coleoptile emergence stage (PO:0007045)- OBO_SF2_PO:370
OBO_SF2_PO: 366
seed imbibition stage (PO:0007022)
seedling development stage (PO:0007131) 
seedling shoot emergence stage (PO:0007130)

Revised name and definition: 
Was:
0 seed germination stage (PO:0007057): The resumption of growth by the embryo in a seed.

Revised: 
seed germination stage (PO:0007057): A sporophyte vegetative stage (PO:0007134) which begins with the onset of a seed imbibition stage (PO:0007022) and ends with a radicle emergence stage (PO: 0007015), or a coleorhiza emergence stage (PO:0025475), as a result of the growth by the plant embryo (PO:0004537) in a seed (PO:0009010).

Added dbxref to  GO:0009845 seed germination: The physiological and developmental changes that occur in a seed commencing with water uptake (imbibition) and terminating with the elongation of the embryonic axis.