Modified definition of protoderm:
protoderm (PO:0006210): 
Was: A portion of meristem tissue (PO:0009013) that develops from the outer layer of an apical meristem (PO:0020144) and gives rise to a portion of epidermis (PO:0005679).

New: A portion of meristem tissue (PO:0009013) that is the outer layer of an apical meristem (PO:0020144), or the outer layer of a plant embryo (PO:0009009), and gives rise to a portion of epidermis (PO:0005679).

Comment: Protoderm may or may not arise from independent initial cells (PO:0004011). Some researchers also call it an epidermis in a meristematic state.

Added new exact synonym: portion of protoderm tissue


adaxial protoderm (PO:0006083): 
was: A portion of protoderm (PO:0006210) tissue that is the outermost layer of the shoot apical meristem (PO:0020148) and gives rise to the leaf adaxial epidermis (PO:0006018).

Added new exact synonym: portion of adaxial protoderm tissue


abaxial protoderm (PO:0006082): 
was: A portion of protoderm (PO:0006210) tissue that is the outermost layer of the shoot apical meristem (PO:0020148) and gives rise to the leaf abaxial epidermis (PO:0006019).

Added new exact synonym: portion of abaxial protoderm tissue