Fixed typo in plant embryo true leaf formation stage (PO:0001095).
Revised def'n: 
plant embryo dormant stage (PO:0025377): 
Was: A sporophyte dormant stage during which a plant embryo participates in a dormancy process. 
New: A mature plant embryo stage (PO:0001081) during the interval between when a plant embryo (PO:0009009) begins a dormancy process (GO:0022611) and one of the following: the beginning of the seed germination stage (PO:0007057) in seed plants, formation of the first vascular leaf (PO:0009025) in pteridophytes, the beginning of development of a sporangium (PO:0025094) in bryophytes, or the beginning of the formation of a plant organ (PO:0009008) such as a root (PO:0009005), shoot axis (PO:0025029), or vascular leaf (PO:0009025) in a cultured plant embryo (PO:0000010).

revised comment: During a plant embryo dormant stage, a plant embryo participates in a dormancy process, but not all plant embryos have a dormant stage. A dormancy process (GO:0022611) is a developmental process in which dormancy (sometimes called a dormant state) is induced, maintained or broken. Dormancy is a suspension of most physiological activity and growth that can be reactivated. In seed plants, a plant embryo dormant stage may be part of a seed dormant stage.

Updated Xref to OBO_SF2_PO:426