paleal apiculus (PO:0006033)- Revised name and def'n: Revised name: palea apiculus Was: A short, abrupt, flexible point present at the apex of the palea.
Revised def'n: A cardinal organ part (PO:0025001) which is a short, abrupt, flexible point at the apex of the palea (PO:0009038).

New term: lemma apiculus (PO:0025616): A cardinal organ part (PO:0025001) which is a short, abrupt, flexible point at the apex of the lemma (PO:0009037).

Revised definitions slightly:

palea (PO:0009038)
Was: A flower bract (PO:0009034) that is the upper/distal of the two bracts (PO:0009055) enclosing a (PO:0009082) spikelet floret.

Revised:  The upper (distal), usually smaller, of the pair of flower bracts (PO:0009034) enclosing a (PO:0009082) spikelet floret.

lemma (PO:0009038):
Was: A flower bract (PO:0009034) that is the lower of the two bracts (PO:0009055) enclosing a (PO:0009082) spikelet flore
Revised: The lower, usually larger, of the pair of flower bracts (PO:0009034) enclosing a (PO:0009082) spikelet floret.