Revised: style (PO:0009074): Current def'n: An elongated part of a carpel or group of fused carpels between the ovary and the stigma, and through which the pollen tube grows.
Revised def'n: A cardinal organ part (PO:0025001) which is an elongated part of a carpel (PO:0009030) or gynoecium (PO:0009062) between the plant ovary (PO:0009072) and the stigma (PO:0009073), and through which the pollen tube (GO:0090406) grows.

Revised comment:
If you are annotating to this structure for Zea mays or other grasses, please also add an annotation to the corresponding floret type. Choose the most specific term possible from: spikelet floret (PO:0009082), ear floret (PO:0006354), upper floret of pedicellate spikelet of ear (PO:0006350), upper floret of sessile spikelet of ear (PO:0006352).

Removed silk as a synonym and the alt id: PO:0006488