pneumatophore (PO:0025357) Revised def'n and comment: was: "A lateral root that is erect and protrudes above the soil or water."
Comment: Pneumatophores are found in trees that live in flooded habitats such as mangroves and often have many lenticels. May provide oxygen to below ground roots growing in flooded soils. Sometime prop roots (PO:0000044) may act as breathing roots. In that case, the annotation should go to PO:0000044, whereas this term should be used for breathing roots arising at or below the soil or water surface. Knee roots are a type of pneumatophore in which the lateral root first grows upward, then turns and grows back down into the substrate, forming a knee-like bend.

Revised def'n:  A lateral root (PO:0020121) that is erect and protrudes above the surface of the soil or water, and may provide oxygen to below-ground roots.

Revised Comment:
Pneumatophores are found in trees that live in flooded habitats such as mangroves, and often have many lenticels (PO:0000031) which may provide oxygen to below-ground roots growing in flooded soils. Knee roots are a type of pneumatophore in which the lateral root first grows upward, then turns and grows back down into the substrate, forming a knee-like bend. Note that a pneumatophore differs from a prop root (PO:0000044), which is formed at a stem node above the ground, and grows downward into the soil.

prop root (PO:0000044):
Revised def'n:  A shoot-borne nodal root (PO:0003005) formed at a stem node (PO:0020141) above the ground and growing directly into the soil.
Revised comment:
This term is perhaps best defined functionally to prop the plant up, but sometimes prop roots may also act as breathing roots. Compare to pneumatophore (PO:0025357), which is a lateral root that grows upwards from below the soil or water surface.