Revised def'n: silk scar  PO:0002004 was: The silk scar is composed of the shrivelled remains of the intercalary region of the silk, together with the base of the silk which is the true stylar tissue of the silk arising from the distal end of the developing kernel (caryopsis).
Currently is_a fruit distal end (PO:0008001), moved to be a cardinal part of multi-tissue plant structure (PO:0025498)

Revised def'n: A cardinal part of multi-tissue plant structure (PO:0025498) which is composed of the shrivelled remains of the silk (PO:0006488), and the base of the style (PO:0009074), and is located on the distal end of a developing or mature kernel (caryopsis fruit; PO:0030104).
relations- part_of pericarp, fruit distal end (PO:0008001)

silk PO:0006488
- Added new comment: During pollination (GO:0009856), the pollen (PO:0025281) grows down through a portion of  transmitting tissue (PO:0020102) to a plant ovary ovule (PO:0025490) and a megagametophyte (PO:0025279) to effect fertilization (GO:0009566).

Renamed parts of plant ovary to include 'plant" as in the parent term:
- ovary locule (PO:0025266)  >> plant ovary locule (PO:0025266)

Was: A locule that is a cavity in an ovary.
Revised def'n:  A locule (PO:0025263) that is a cavity in a plant ovary (PO:0009072).

Other locules:
fruit locule (PO:0025270): A locule (PO:0025263) that is a cavity in a fruit (PO:0009001) and develops from a plant ovary locule (PO:0025266).

sporangium locule (PO:0025265):  A locule (PO:0025263) that is a cavity in a sporangium (PO:0025094) or several fused sporangia.

ovary wall (PO:0005022) >> plant ovary wall (PO:0005022)
Was: A wall of an ovary that is composed of abaxial (outer) and adaxial (inner) epidermis and a middle layer, composed of mesophyll cells and a distinct subepidermal adaxial cell layer (as found in Arabidopsis).

A portion of plant tissue (PO:0009007) that is composed of plant ovary abaxial (outer) epidermis (PO:0005023), plant ovary adaxial (inner) epidermis (PO:0005024), and a plant ovary wall middle layer (PO:0005025).