WAS: abaxial epidermis of ovary (PO:0005023): The epidermal cell layer of the abaxial/outer surface of the ovary. REVISED: plant ovary abaxial epidermis (PO:0005023): A portion of ovary epidermis (PO:0006063) which is the outer surface of a plant ovary (PO:0009072).
WAS:  adaxial epidermis of ovary  (PO:0005024)
REVISED:  plant ovary adaxial epidermis (PO:0005024): A portion of ovary epidermis (PO:0006063) which is the inner surface of a plant ovary (PO:0009072).

WAS: middle layer of ovary wall (PO:0005025): A portion of tissue that is composed of one or more mesophyll layers. It may have a distinct subepidermal adaxial cell layer (as found in <i>Arabidopsis</i>).
REVISED: plant ovary wall middle layer (PO:0005025): A portion of chlorenchyma tissue (PO:0005426) that is part of the plant ovary wall (PO:0005022) and is composed of mesophyll cells (PO:0004006).

Revised comment: In some species, such as Arabidopsis, there may be a distinct subepidermal adaxial cell layer. The cells in the mesophyll layer(s) may not differentiate similar to the spongy and palisade cells found in the leaf mesophyll.